requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.3
django-import-export==4.4.0
django-simple-history==3.11.0
sqlparse==0.5.5
tablib==3.9.0
